nullFields
